Fixes crash when opening the app from background (#50)
Problem: I see some random crashes in GoBot code regarding exc_bad_access.
This kind of errors are hard to spot because are most of the time related
to multrithreading issues. I found that when opening the app from background
we call resume to gobot, but inside we don't use the gobot thread, this should
be causing memory issues a bit later.

Solution: Use the thread inside the appcycle functions.