Move UITextView for post editing into a custom view (#78)
The plain UITextView that is currently used to compose a post now lives
in a custom UIView. The custom view also offers an "attributedText"
property that acts just like the one from the original UITextView. It
always holds the Markdown source including attributed mention text.

However this view also supports switching between the "compose" view
with the Markdown source and a non-editable preview of the rendered
post.

The button for switching the view is going to be added in the next
commits. At this point the change should be invisible to the user.