Add Markdown preview during post creation (#195)
* Move UITextView for post editing into a custom view (#78)

The plain UITextView that is currently used to compose a post now lives
in a custom UIView. The custom view also offers an "attributedText"
property that acts just like the one from the original UITextView. It
always holds the Markdown source including attributed mention text.

However this view also supports switching between the "compose" view
with the Markdown source and a non-editable preview of the rendered
post.

The button for switching the view is going to be added in the next
commits. At this point the change should be invisible to the user.

* Add a post preview button to NewPostViewController (#78)

The button switches the new post composer view between editing and
preview mode.

I am not sure if this is the most aesthetic and usable way to add this
button. I thought about using an SF Symbols icon instead though *only* an
icon isn't really self explanatory.
Besides, for screen space reasons the "Markdown supported" label had to
go. Therefore a different means is needed to communicate Markdown
capability. On one hand, aving just "Preview" as button caption could mean
anything. Saying "Preview Markdown" on the other hand makes the button
waaay too wide.

* Extend MarkdownTextView to animate the preview (#78)

Using two separate UITextViews for source and rendered text, the view
animates between the two views instead of exchanging the UITextView
contents directly.

* Change the Markdown preview button to a UISwitch

...and adjust the "Markdown supported" label to say "Markdown preview".
It now layouts against the preview toggle like a caption. This
implements the design suggested by @Chardot.

* Speed up markdown editing/preview switch animation

The animation duration now matches the speed of the UISwitch more
closely when toggling it.

* Add Markdown preview to replies

Update ReplyTextView to match strategy made when creating posts.
ReplyTextView contains two text views just like PostTextEditorView.

* Fix compile error related to preview toggle

The previous commit used the old preview button, this one fixes
that and uses the preview toggle when replying to a post.

* Style New Post Editor background color and switch

Update the color of the New Post editor so it has more constrast
when editing, and the color of the switch so it matches the branding.

* Disable New Post editor when previewing

Disable editing when displaying the Markdown preview because
any change made there will be lost anyway.

Co-authored-by: Martin Dutra <martindsq@hotmail.com>