Accept all peer connections
A while ago a mechanism which tries to filter accepted connections was
introduced in go-ssb. By defaut this mechanism accepts connections based on the
follow graph. This may make sense if go-ssb acts as a pub but makes little
sense if go-ssb runs on a phone as your social media profile node. After all as
a user I want people to connect to me and get my messages. My profile is
public.

On top of all of that the feature doesn't seem to work correctly. Even after I
followed my Patchwork identity the connections from/to Patchwork were still
being rejected.

This sets an authorizer which accepts all peer connections to mitigate this
problem.

Fixes #431.