Change C.longlong to the correct type
I don't fully understand why we used a cast to C.longlong instead of just
casting to C.int64_t as those types are not always the same. I think it is safe
to change this so that it compiles on both MacOS and GNU/Linux or as I have
recently taken to calling it GNU+Linux.