Make builds (a bit more)reproducible
It has been annoying me for a while that even without any changes to the code
builds change the .a files and I can't tell if the build was needed or not.

I right away discovered that the problem is related to timestamps thanks to
running vimdiff on xxd output. I was trying to liberally apply faketime without
any success for hours. Don't ask me for how long exactly because it is
embarassing. In the end I was helped out a ton and told that the problem was
apparently related to the modification times in the ar archive. Before I
started this stupid adventure I didn't even know that ar, libtool or even
something as simple as diffoscope exist.

>Thanks /u/Foxboron for this information.